backend/requirements-dev.txt
# 개발용 의존성
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.17.2
black>=23.11.0
flake8>=6.1.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
watchdog>=3.0.0
